Move Themes under dashboard and add access restrictions
This change
* Moves the themes controller under the Admin namespace
* Updates the routing to `/admin/themes`
* Adds a link to themes in the Admin sidebar
* Adds access restricitons to prevent unauthorized users from managing themes
* Grants access to `admin/themes/current.css` for any user or guest

Access to the current theme css is required to render the desired
colors and styling for all users regardless of their login status.
