+++ app/views/admin/themes/_theme.html.erb
@@ -1,0 +1,27 @@
+<div id="<%= dom_id theme %>" class="theme">
+
+  <div class="canvas">
+    <!-- Label-->
+    <div class="theme-label">
+      <div class="field-label visually-hidden">Label:</div>
+      <div class="field-value"><h2><%= theme.label %></h2></div>
+    </div>
+
+    <!-- Loading spinner -->
+    <div class="lds-dual-ring"></div>
+
+    <!-- Site preview -->
+    <div class="preview-container">
+      <div class="preview">
+        <% if theme.persisted? %>
+          <iframe src="/catalog?theme_id=<%= theme.id.to_s -%>"></iframe>
+        <% else %>
+          <%= image_tag 'big_plus.png' %>
+        <% end %>
+      </div>
+    </div>
+
+    <!-- Control buttons -->
+    <%= render(partial: 'theme_controls', locals: {theme: theme}) if theme %>
+  </div>
+</div>
