+++ app/views/admin/themes/_theme.json.jbuilder
@@ -1,0 +1,3 @@
+json.extract! theme, :id, :label, :site_name, :header_color, :header_text_color, :background_color,
+              :background_accent_color, :created_at, :updated_at
+json.url theme_url(theme, format: :json)
