+++ app/views/admin/themes/_theme_controls.html.erb
@@ -1,0 +1,20 @@
+<div class="theme-controls">
+  <% case action_name %>
+  <% when 'index' %>
+    <% if theme.persisted? %>
+      <%= link_to "Preview", theme, id: dom_id(theme, :preview), class: 'btn btn-primary btn-sm' %>
+    <% else %>
+      <%= link_to "New Theme", new_theme_path, id: 'new_theme', class: 'btn btn-primary btn-sm' %>
+    <% end %>
+  <% when 'show' %>
+    <% if @theme.active %>
+      <%= button_to "Active", '#', id: 'activate-theme', class: 'btn btn-outline-success', disabled: true %>
+    <% else %>
+      <%= button_to "Make Active", activate_theme_path(theme), method: :patch, id: 'activate-theme', class: 'btn btn-success' %>
+    <% end %>
+    <%= link_to "Edit", edit_theme_path(theme), id: 'edit-theme', class: 'btn btn-primary' %>
+    <%= link_to "Back to themes", themes_path, id: 'list-themes', class: 'btn btn-primary' %>
+    &nbsp; &mdash; &middot; &mdash; &nbsp;
+    <%= button_to "Delete", theme, method: :delete, id: 'delete-theme', class: 'btn btn-danger' %>
+  <% end %>
+</div>
