+++ app/views/admin/themes/edit.html.erb
@@ -1,0 +1,10 @@
+<h1>Editing theme</h1>
+
+<%= render "form", theme: @theme %>
+
+<br>
+
+<div>
+  <%= link_to "Show this theme", @theme %> |
+  <%= link_to "Back to themes", themes_path %>
+</div>
