+++ app/views/admin/themes/index.json.jbuilder
@@ -1,0 +1,1 @@
+json.array! @themes, partial: 'themes/theme', as: :theme
