+++ app/views/admin/themes/new.html.erb
@@ -1,0 +1,9 @@
+<h1>New theme</h1>
+
+<%= render "form", theme: @theme %>
+
+<br>
+
+<div>
+  <%= link_to "Back to themes", themes_path %>
+</div>
