+++ app/views/admin/themes/show.css.erb
@@ -1,0 +1,12 @@
+:root {
+/*  Consider settings for these https://getbootstrap.com/docs/5.3/customize/color-modes/  */
+  --header-background:  <%= @theme.header_color %>;
+  --header-search-background:  <%= @theme.background_accent_color %>;
+  --footer-background:  <%= @theme.header_color %>;
+  --footer-text:        <%= @theme.header_text_color %>;
+  --main-background:    <%= @theme.background_color %>;
+  --bs-dark-rgb:        <%= rgb(@theme.header_color) %>;
+  --bs-navbar-color:    <%= rgb(@theme.header_text_color) %>;
+  --bs-body-bg:         <%= @theme.background_color %>;
+  --bs-body-bg-rgb:     <%= rgb(@theme.background_color) %>;
+}
