+++ app/views/admin/themes/show.html.erb
@@ -1,0 +1,3 @@
+<h1>Theme Preview</h1>
+
+<%= render partial: 'theme', object: @theme %>
