+++ app/views/admin/themes/show.json.jbuilder
@@ -1,0 +1,1 @@
+json.partial! 'themes/theme', theme: @theme
