+++ spec/views/admin/themes/show.html.erb_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe 'themes/show' do
+RSpec.describe 'admin/themes/show' do
   let(:theme)  { FactoryBot.create(:theme, active: false) }
 
   before do
